Re-organize tests
The version annotations can be per method,
so everything can be moved to one class.
