+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r61/InvalidateVirtualFileSystemCrossVersionSpec.groovy
@@ -16,21 +16,61 @@
 
 package org.gradle.integtests.tooling.r61
 
+import org.gradle.integtests.fixtures.daemon.DaemonFixture
 import org.gradle.integtests.fixtures.executer.GradleExecuter
 import org.gradle.integtests.tooling.fixture.TargetGradleVersion
+import org.gradle.integtests.tooling.fixture.ToolingApiSpecification
+import org.gradle.integtests.tooling.fixture.ToolingApiVersion
 import org.gradle.test.fixtures.server.http.BlockingHttpServer
 import org.gradle.tooling.model.gradle.GradleBuild
 import org.junit.Rule
 
-@TargetGradleVersion(">=6.1")
-class InvalidateVirtualFileSystemCrossVersionSpec extends AbstractInvalidateVirtualFileSystemCrossVersionSpec {
+import java.nio.file.Path
+import java.nio.file.Paths
+
+@ToolingApiVersion(">=6.1")
+class InvalidateVirtualFileSystemCrossVersionSpec extends ToolingApiSpecification {
     @Rule BlockingHttpServer server = new BlockingHttpServer()
     GradleExecuter executer
 
+    List<String> changedPaths = [file("src/main/java").absolutePath]
+
     def setup() {
+        toolingApi.requireIsolatedToolingApi()
+
+        buildFile << """
+            apply plugin: 'java-library'
+        """
+
         executer = toolingApi.createExecuter()
     }
 
+    @TargetGradleVersion(">=2.6")
+    def "no daemon is started for request"() {
+        when:
+        withConnection { connection ->
+            connection.notifyDaemonsAboutChangedPaths(toPaths(changedPaths))
+        }
+
+        then:
+        toolingApi.daemons.daemons.empty
+    }
+
+    @TargetGradleVersion(">=6.1")
+    def "changed paths need to be absolute"() {
+        changedPaths = ["some/relative/path"]
+
+        when:
+        withConnection { connection ->
+            connection.notifyDaemonsAboutChangedPaths(toPaths(changedPaths))
+        }
+
+        then:
+        def exception = thrown(IllegalArgumentException)
+        exception.message == "Changed path '${Paths.get(changedPaths[0])}' is not absolute"
+    }
+
+    @TargetGradleVersion(">=6.1")
     def "invalidates paths for single idle daemon"() {
         when:
         createIdleDaemon()
@@ -43,6 +83,7 @@ class InvalidateVirtualFileSystemCrossVersionSpec extends AbstractInvalidateVirt
         pathsInvalidated()
     }
 
+    @TargetGradleVersion(">=6.1")
     def "does not invalidate paths for single busy daemon"() {
         server.start()
         buildFile << """
@@ -72,10 +113,49 @@ class InvalidateVirtualFileSystemCrossVersionSpec extends AbstractInvalidateVirt
         build?.waitForFinish()
     }
 
+    @TargetGradleVersion(">=2.6 <6.1")
+    def "invalidating paths has no effect"() {
+        when:
+        createIdleDaemon()
+        def executedCommandCount = countExecutedCommands()
+
+        then:
+        executedCommandCount == 3
+
+        when:
+        withConnection { connection ->
+            connection.notifyDaemonsAboutChangedPaths(toPaths(changedPaths))
+        }
+
+        then:
+        !pathsInvalidated()
+        countExecutedCommands() == executedCommandCount
+    }
+
+    private int countExecutedCommands() {
+        toolingApi.daemons.daemon.log.count("command = ")
+    }
+
     private void createIdleDaemon() {
         withConnection { connection ->
             connection.model(GradleBuild).get()
         }
         toolingApi.daemons.daemon.assertIdle()
+    }
+
+    def cleanup() {
+        toolingApi.close()
+    }
+
+    boolean pathsInvalidated() {
+        pathsInvalidatedForDaemon(toolingApi.daemons.daemon)
+    }
+
+    boolean pathsInvalidatedForDaemon(DaemonFixture daemon) {
+        daemon.log.contains("Invalidating ${changedPaths}")
+    }
+
+    static List<Path> toPaths(List<String> paths) {
+        paths.collect { Paths.get(it) }
     }
 }
